Ensure StyleSheetInsert-bootstrap.html doesn't spam willInsertBody IPCs
A recent change to support ResourceScheduler in OOPIF mode [1] caused
a regression on this benchmark. The reason is that this benchmark
Essentially creates an empty iframe, populates it, and empties it again
in a tight loop.

This triggers us to send a WillInsertBody IPC for every loop iteration,
adding a lot of overhead to the benchmark.

As this idiom is likely not representative, this patch modifies the
benchmark to instead just insert and remove the style node in a tight
loop.

[1] https://codereview.chromium.org/2655393004

BUG=689190

Review-Url: https://codereview.chromium.org/2684483002
Cr-Original-Commit-Position: refs/heads/master@{#448621}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d68b4a106ecc41b8b16079a30ac422ad571554ed