Refactor Samsung SmartClip implementation.
Samsung "Smart Clip" (a.k.a. "Smart Select") is a Galaxy Note-specific
proprietary API to select a rectangle with the stylus and extract the text
contents.  This moves it from the ContentViewCore->ViewMsg->WebViewImpl
flow into the more modern WebContents->RenderFrameHost->WebLocalFrame
path.

Notes:
- One of the values we're providing in the return data is the original
rectangle that was passed into us.  It's currently being passed in back
and forth from Blink unnecessarily, and wasn't DIP-adjusted properly.
Empirically on a Galaxy Note 5, there's no evidence that the OEM code
ever looks at this rectangle, so just delete it.

- This never supported subframes (even local iframes) and still doesn't.
This patch is intended as a no-op refactoring for now; for subframe
support, we might consider generating this out of the accessibility
trees in the future.

BUG=665665
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2667283007
Cr-Original-Commit-Position: refs/heads/master@{#448816}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3d72e6945823292c6fa7487268333b538b6a10e