PaymentApp: Implement invokePaymentApp() in renderer side.
This is the final patch of a series of three that implements invokePaymentApp().
  - 1/2: https://codereview.chromium.org/2610163002/ (browser side)
  - 2/2: This patch (renderer side)

After this patch, it is possible to invoke a selected payment app if calling
the invokePaymentApp() method in PaymentRequest UI implementation. This CL also
includes a test to verify such behavior.

BUG=661608
TEST=payment_app_browsertest.cc

Review-Url: https://codereview.chromium.org/2646313002
Cr-Original-Commit-Position: refs/heads/master@{#448906}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4545cd9e65fd3a2276e1676ed57c85d1bb7631a