URL.search should return the set value via setSearch
DOMURLUtils and KURL trim off the leading question mark independently, which
leads to trimming off two leading questions in addition.
DOMURLUtilsReadOnly::search recovers one question mark and we lose one question
mark in total.

The special handling in DOMURLUtils is needed because KURL::setQuery handles
an empty string and a null string differently. KURL::setQuery("?") is
equivalent to KURL::setQuery("") but what we need is KURL::setQuery(String()).

This CL fixes the problem by calling KURL::setQuery(String()) if the given
value is "?", and calling KURL::setQuery with the given string if it starts
with a question mark and consists of two or more characters.

BUG=682150

Review-Url: https://codereview.chromium.org/2686653003
Cr-Original-Commit-Position: refs/heads/master@{#448940}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29fc3042b1b3bea3e71a2c5b0a9a81ce435c2d39