Remove GraphicsLayer::didScroll and directly call ScrollableArea::didScroll
This patch removes GraphicsLayer::didScroll, replacing this cc->blink
scroll callback with a direct call to Scrollablearea::didScroll. The
old function on GraphicsLayer was only used to lookup the scroll
offset but we can pass that along in the didScroll callback.

This patch will be useful for SPV2 which does not use GraphicsLayer and
can benefit from avoiding it, but this patch is intended to be useful
for SPV1. This code has been split off of a larger patch for composited
scrolling: https://codereview.chromium.org/2676923002.

BUG=667946
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2680953002
Cr-Original-Commit-Position: refs/heads/master@{#449073}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 27a8cfde871ad4786943a345569f36c9a5ba6e29