EME: Allow case-insensitive parameter names in media MIME types
According to RFC 2045,  "All media type values, subtype values, and
parameter names as defined are case-insensitive." Switch
requestMediaKeySystemAccess() to use ContentType, which is not
case-sensitive.

Added new test that should fail. Prior to this change it would pass as
the CODECS="foo" would be ignored, and it would look like a container
only MIME type, which is allowed.

BUG=605661
TEST=new test fails as expected

Review-Url: https://codereview.chromium.org/2681993002
Cr-Original-Commit-Position: refs/heads/master@{#449169}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5784d408a464915895f9b47385df377a555032e