Add DCHECK in copy constructor of StyleVisualData for m_zoom.
Currently StyleVisualData::m_zoom is reset to the initial zoom value
every time StyleVisualData is copied using the copy constructor. There
is no indication that this is intentional.

This patch adds a DCHECK to see if there's any case where we trigger
this bug. If not, we will copy m_zoom properly in a separate patch.

Not to be merged in.

BUG=689330

Review-Url: https://codereview.chromium.org/2687533004
Cr-Original-Commit-Position: refs/heads/master@{#449215}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4be275ae7e189294e00a1c65ec9976974b1329de