Temporary disable DCHECK for transform ancestor checking in CompositedLayerMapping
This patch disables |DCHECK()| fro transform ancestor checking in
|CompositedLayerMapping| for landing [1] and [2].

As of offline discussion, this DCHECK itself can be safely ignored since there
is no stability or security implication, and doesn't result in rendering glitch
in this particular case. However, it is hard to make minimal reproduce case;
this DCHECK hits on |LoginTest.GaiaAuthOffline| which is run on ChromeOS only.

[1] http://crrev.com/2680943004 Make FrameSelection to hold non-canonicalized positions
[2] http://crrev.com/2340463002 Set default focus on offline signin page

BUG=646437
TEST=n/a; no behavior changes

Review-Url: https://codereview.chromium.org/2687043002
Cr-Original-Commit-Position: refs/heads/master@{#449539}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a22b0b4b1401f49e388eef9f96d5cf22e2b5fa08