WPT Auto-import: fix for owner list in import CLs.
This is a follow-up fix for http://crrev.com/2684663003.

The mistake was: In one part of the code I was assuming that
get_directory_owners() would return a map from owner to list of
directories, but in reality it returned a dict from owner to single
directory (string).

We want to have lists of directories, since a single import CL
will often have modifications to many directories for one owner.

This resulted in silly-looking CL descriptions like
http://crrev.com/2687343002.

BUG=687270

Review-Url: https://codereview.chromium.org/2684423004
Cr-Original-Commit-Position: refs/heads/master@{#449666}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ccb652d00c6fbd072608a7bdbb9f1fbe17fd87d