[LayoutNG] Implement very simple line breaker
This patch implements a very simple line breaker.

This line breaker is for the proof of design of NGLineBuilder. The plan
is to replace this with the shaper-driven line breaker.

Changes in this patch include:
* Allow lines to break at the mid of an item.
  - NGLineBuilder accepts index and offset in its functions.
  - NGPhysicalTextFragment has index, start_offset, and end_offset.
* A new set of APIs in NGLineBuilder:
  - APIs are changed to range (start/end). With previous AddItem() APIs,
    callers must keep track of the last item, and items were DCHECKed to
    be contiguous. Range-based APIs simplifies both NGLineBuilder and
    its callers.
  - Add break opportunity and hangable for the better usability.
* NGLayoutInlineItem::InlineSize() supports start/end offset.
  - This may be removed with the shaper-driven line breaker.

The internal structure of NGLineBuilder isn't very much updated with the
API changes. This will be reviewed again once the APIs are more
finalized.

BUG=636993

Review-Url: https://codereview.chromium.org/2682833004
Cr-Original-Commit-Position: refs/heads/master@{#449856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 01199c020ca74ac8d32a7d2939d0a090cf08d801