ColumnBalancer: need to examine lines that protrude into the flow thread portion.
We cannot skip a line that starts before the flow thread portion of interest,
if it ends inside the portion.

Some extra care is needed to avoid regressions here: The part of a line that
starts before the multicol container itself needs to be ignored, or we risk
overstretching the multicol container.

BUG=642181

Review-Url: https://codereview.chromium.org/2690863003
Cr-Original-Commit-Position: refs/heads/master@{#450033}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e6930336491fd5bc277dc9476873a684ebeae44