Re-enable DCHECK for pointertypes of coalesced events
The modified DCHECK was failing on MAC when mouse was
leaving window due to the missing pointer type from
the event Chrome was receiveing from the OS. So we
were setting the pointerType to Unknown.
After the following fix:
https://chromium.googlesource.com/chromium/src/+/e97eb52977dd7ed3936227f0d240cfeed865e520
we have a heuristic in place to set the type to either
pen or mouse. So this DCHECK can be enabled again.

BUG=684292

Review-Url: https://codereview.chromium.org/2688793003
Cr-Original-Commit-Position: refs/heads/master@{#450075}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b4cb83ed1e58193e8f3935ff406ce0ddb6310c6