Add Jinja2 macro for adding trailing strings.
There's a lot of repetitive code in ComputedStyleBase.h.tmpl to handle
formatting. For example, the member initialiser list in the generated
constructor needs a trailing comma for every member except the last one.
This patch moves such code into a generic macro in macros.tmpl, which
which makes it available for other templates.

This will be used more frequently when we start introducing more
complexity to the ComputedStyle templates.

BUG=628043

Review-Url: https://codereview.chromium.org/2680123002
Cr-Original-Commit-Position: refs/heads/master@{#450173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c0a8bdbcd693f00a522afa34e5bdceb3129bb1da