Changed EPosition to an enum class and renamed its members
Changed EPosition to an enum class and gave it an unsigned underlying
type. Also renamed its members to match the CSS keywords they
represent, and added the 'k' prefix in line with the Blink rename to
match Chromium style.

Changing it to an enum class enforces better namespacing and code
practices. Adding the unsigned underlying type is pre-work for when the
class is eventually stored as an enum bitfield (it would be done in this
patch, except a presubmit warning already exists that prevents that. The
presubmit warning needs to be updated before that change can occur.)

This is also pre-work to allow EPosition to be used by generated
fields in ComputedStyleBase.

BUG=684966
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2667493002
Cr-Original-Commit-Position: refs/heads/master@{#450202}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 30aef3be9b857563cfab70278e011dba9d1e8a14