UseCounter: Introduce UseCounter::Observer for layout tests
This is a follow-up CL for [1]. We added some layout tests in the CL. They
assume MessagePort and UseCounter share the same IPC channel and their messages
are handled in order. However, this assumption will be untrue by MessagePort
mojofication[2] and the layout tests will be flaky.

To fix this, this CL stops assuming the message order and instead introduces a
mechanism to observe UseCounter changes.

[1] https://codereview.chromium.org/2586863002/
[2] https://codereview.chromium.org/2422793002/

BUG=376039

Review-Url: https://codereview.chromium.org/2680423006
Cr-Original-Commit-Position: refs/heads/master@{#450241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 452a477ff34a2af79a882b44ccc1939230c9e527