Remove unnecessary refreshCastButtonVisiblity call
computeWhichControlsFit is being called on every mousemove event on a video
(via a call to refreshCastButtonVisiblity). After discussing offline with
mlamouri@ and avayvod@, it was determined that the call to
refreshCastButtonVisiblity is no longer needed, and therefore can be removed.

In addition, this CL refactors some media layout tests that test the cast button visibility, and appends to an existing cast button visibility test (controls-cast-button-narrow.html) to test for correct visibility after expanding and contracting a video.

BUG=639273

Review-Url: https://codereview.chromium.org/2684973004
Cr-Original-Commit-Position: refs/heads/master@{#450524}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fb1451de0171f8bdd0915c22016750d79c0b3ac