webkitpy: Reworking test chunking for sharding.
Tests which are not being run as part of this shard should not appear in
the output. This means the chunking needs to occur *before* the test
expectations are loaded as any test with an expectation will appear in
the output.

After doing this;

 * It is possible to merge the test shard output without having to
   understand the different test states.

 * It significantly reduces the size of the output on each shard (as
   they only contain the details about tests which are actually assigned
   to the shard).

BUG=524758

Review-Url: https://codereview.chromium.org/2693533003
Cr-Original-Commit-Position: refs/heads/master@{#450623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 238f6ccccaef8bff5a9f2be96ce71bdf2faebde6