CSP: Add test: Allow port upgrade without scheme upgrade and vice versa.
This is a backport of a test in:
https://codereview.chromium.org/2612793002/

What is interesting in this test is that CSP allow port and scheme to be
upgraded in their secure version independently. The result is that one
of them can be upgraded without the other (http on port 443 or https on
port 80).

BUG=692499

Review-Url: https://codereview.chromium.org/2689363003
Cr-Original-Commit-Position: refs/heads/master@{#450685}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ba1205f748c335e037ad516e29cee9d23dffd21f