Recognize skbug.com bugs in TestExpectations
Changes in Skia often cause changes to layout tests. In some cases
there is a pre-existing Skia issue open for the fix which should be
used, but currently cannot. The procedure up to now has been to create
a Chromium bug just for the rebaseline. It would be nice to instead
associate the rebaseline with the original issue in Skia.

BUG=skia:6203

Review-Url: https://codereview.chromium.org/2695273002
Cr-Original-Commit-Position: refs/heads/master@{#450777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80fa510831e8ebf354e79a24249cb6e8428df6cb